refactor(scanner): Move `ScanResult.filterByVcsPath()` to utils
Prepare for re-use in an upcoming change.
